requirements.txt
alembic==1.13.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
dnspython==2.6.1
docker==7.1.0
eventlet==0.35.1
Flask==3.1.1
-e git+https://github.com/corlukantonio/Flask-JSGlue.git@e49d34d6f6ff88cc5ccb74f25914c9166439e7cc#egg=Flask_JSGlue
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.1
pytest-env==1.1.3
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
SQLAlchemy==2.0.27
testcontainers==3.7.1
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
